Added multi-cylinder

+++ includes/utils/turtle.hpp
@@ -43,7 +43,7 @@ public:
                          const std::vector<Vector3> &pts,
                          const std::vector<std::vector<size_t>> &faces);
     // convert edges to cylinder
-    void to_cylinder(const size_t &radius, const size_t &discretisation,std::vector<Vector3> &pts,
+    void to_cylinder(const double &radius, const size_t &discretisation,std::vector<Vector3> &pts,
                      std::vector<std::vector<size_t>> &faces);
     std::vector<Vector3> compute(const core::State &sentence, const std::string &ignore,
             std::vector<std::vector<size_t>> &faces,
